+++ CHANGELOG.md
@@ -5,6 +5,19 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
+## Unreleased
+
+### Changed
+- Updated JUnit from 4.13.2 to 5.9.1 (tests)
+- Updated grel-functions-java from v0.7.3 to v0.9.0 (tests)
+- Updated jena-arq from 3.17.0 to 4.6.1
+- Updated Lombok from 1.18.22 to 1.18.24
+- Updated slf4j from 1.7.36 to 2.0.5
+- Java language version changed to 11; requires JRE >= 11 to run
+
+### Fixed
+- be.ugent.idlab.knows.functions.agent.AgentTest.testGrelClassesOnClassPathRemoteFnODoc: point remote `grel.ttl` to specific GitHub commit to keep function definitions and implementations in sync.
+
 ## [0.2.1] - 2022-10-10
 
 ### Fixed